Add iframe-prod.js and iframe-staging.js
Create two new iframe scripts to send to clients so that they can always
use a specific domain regardless. This is useful if a client only ever
wants to test staging on an iframe (or only prod).

Split out the common iframe js code into iframe-common.js

J=SPR-2510
TEST=manual

Create an iframe_test.html, iframe_prod.html, iframe_staging.html

<html><head></head><body><div id="answers-container"></div><script src="iframe.js"></script></body></html>
<html><head></head><body><div id="answers-container"></div><script src="iframe-prod.js"></script></body></html>
<html><head></head><body><div id="answers-container"></div><script src="iframe-staging.js"></script></body></html>

Test to see iframe url generated on `grunt webpack` is correctly
pointing to prod, prod, staging respectively.

(export JAMBO_INJECTED_DATA='{"businessId":<redacted>,"answers":{"experiences":{"verizon_answers":{"apiKey":"<redacted>"}}},"pages":{"stagingDomains":["yextpages.net","landingpagespreview.com"],"domains":{"prod":{"domain":"answers.verizon.com","isHttps":true},"staging":{"domain":"answers_verizon_com.yextpages.net","isHttps":true}}}}'; jambo build && grunt webpack)
